feat: Add news articles to ChatHome component
This commit adds the functionality to display news articles in the ChatHome component. It introduces a new prop `news` to the `ChatHome` component, which accepts an array of `NewsArticleModel` objects. The news articles are fetched using the `FindAllNewsArticles` service.

The `ChatHome` component now renders the news articles section, displaying each article using the `NewsArticle` component. If there are no news articles available, a message indicating so is shown.

This feature enhances the user experience by providing relevant news articles within the chat interface.