Chat with File Preparedness (#121)
* Refactor login error message for better user experience

* feat: Update file upload description in ChatFileUI component

Update the file upload description in the ChatFileUI component to provide more information to users. The updated description now mentions that uploaded files will only be available for a guaranteed seven days. This change is made to align with the new file handling process in ${AI_NAME}.

* feat: Update file upload description in ChatFileUI component

* Update file upload description in ChatFileUI component

* feat: Update file upload description in ChatFileUI component

Update the file upload description in the ChatFileUI component to mention that uploaded files will only be available for a guaranteed seven days. This aligns with the new file handling process in ${AI_NAME}.

* save document content to ChatDocumentModel

* feat: Update app version to 1.5.5 and file handling description in ChatFileUI component

Update the app version to 1.5.5 in the app-global.ts file and package.json. Also, update the file upload description in the ChatFileUI component to mention that uploaded files will only be available for a guaranteed seven days. This aligns with the new file handling process in ${AI_NAME}.

---------

Co-authored-by: Bruno Piovan <bruno.piovan@chde.qld.gov.au>