feat: Add loading components for tenants and pages (#135)
* feat: Add loading components for tenants and pages

* feat: Add "/settings/tenants" to requireAdminPaths in middleware.ts

* feat: Remove unused code and update loading components

Remove unused code related to the "MSSQL" chat type in the chat models. Update the loading components for the chat thread and history pages to use the LoadingSkeleton component.

* feat: Update grid layout in ReportingMessageContainer

Update the grid layout in the ReportingMessageContainer component to have 5 columns instead of 3. This change improves the visual alignment of the elements in the container.

* feat: Update loading components to use PlainLoadingSkeleton

Update the loading components in the settings pages to use the PlainLoadingSkeleton component instead of the LoadingSkeleton component. This change ensures consistent styling and improves the loading experience for users.

* Update loading components to use PlainLoadingSkeleton

* feat: Add dateOnBoarded field to tenant details response

Add the dateOnBoarded field to the tenant details response in the route.ts file. This change ensures that the date of onboarding is included in the response, allowing for better tracking and analysis of tenant data.