Feat/cosmos login (#168)
* switched to azure mi based cosmos auth

* feat: Add COSMOS_DB_ENDPOINT environment variable

The code changes involve adding the COSMOS_DB_ENDPOINT environment variable to the .env.example file and the environment.d.ts file. This variable is required for connecting to the Cosmos DB service. Developers should provide the appropriate value for this variable when setting up the application.

Note: The commit message has been generated based on the provided code changes and recent repository commits.

* chore: Remove javascript code generated by QChat

* refactor: Remove unnecessary code for Cosmos container caching

The code changes involve removing the unnecessary code related to container caching in the `cosmos.ts` file. This code was previously used to store container instances in a cache, but it is no longer needed. This change simplifies the code and improves performance.

Note: The commit message has been generated based on the provided code changes and recent repository commits.

* feat: Update npm dependency to latest stable version

* chore: Update npm lint script to run all linting tasks in parallel

The code changes involve updating the `lint` script in the `package.json` file. Previously, the script used the `npm-run-all` package to run the `lint:js` and `lint:css` tasks in parallel. However, with this change, the script has been updated to use the wildcard `lint:*` to run all linting tasks in parallel. This change simplifies the linting process and improves efficiency.

Note: The commit message has been generated based on the provided code changes and recent repository commits.